+++ internal/app/init/internal/reg/reg.go
@@ -8,7 +8,9 @@ import (
 	"context"
 	"fmt"
 	"io"
+	"os"
 	"path/filepath"
+	"strings"
 	"sync/atomic"
 
 	"github.com/golang/protobuf/ptypes/empty"
@@ -22,6 +24,9 @@ import (
 	"github.com/talos-systems/talos/internal/pkg/upgrade"
 	"github.com/talos-systems/talos/pkg/userdata"
 )
+
+// OSPathSeparator is the string version of the os.PathSeparator
+const OSPathSeparator = string(os.PathSeparator)
 
 // Registrator is the concrete type that implements the factory.Registrator and
 // proto.Init interfaces.
@@ -168,4 +173,66 @@ func (r *Registrator) CopyOut(req *proto.CopyOutRequest, s proto.Init_CopyOutSer
 	}
 
 	return nil
+}
+
+// LS implements the proto.InitServer interface.
+func (r *Registrator) LS(req *proto.LSRequest, s proto.Init_LSServer) error {
+
+	if req == nil {
+		req = new(proto.LSRequest)
+	}
+	if !strings.HasPrefix(req.Root, OSPathSeparator) {
+		// Make sure we use complete paths
+		req.Root = OSPathSeparator + req.Root
+	}
+	req.Root = strings.TrimSuffix(req.Root, OSPathSeparator)
+	if req.Root == "" {
+		req.Root = "/"
+	}
+
+	var maxDepth int
+	if req.Recurse {
+		if req.RecursionDepth == 0 {
+			maxDepth = -1
+		} else {
+			maxDepth = int(req.RecursionDepth)
+		}
+	}
+
+	return filepath.Walk(req.Root, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			// Send errors upstream so that we do not abort the path walk
+			return s.Send(&proto.FileInfo{
+				Name:  path,
+				Error: err.Error(),
+			})
+		}
+
+		err = s.Send(&proto.FileInfo{
+			Name:     path,
+			Size:     info.Size(),
+			Mode:     uint32(info.Mode()),
+			Modified: info.ModTime().Unix(),
+			IsDir:    info.IsDir(),
+		})
+		if err != nil {
+			return err
+		}
+
+		if info.IsDir() && atMaxDepth(maxDepth, req.Root, path) {
+			return filepath.SkipDir
+		}
+		return nil
+	})
+}
+
+func atMaxDepth(max int, root, cur string) bool {
+	if max < 0 {
+		return false
+	}
+	if root == cur {
+		// always recurse the root directory
+		return false
+	}
+	return (strings.Count(cur, OSPathSeparator) - strings.Count(root, OSPathSeparator)) >= max
 }
