+++ internal/app/init/proto/api.proto
@@ -9,6 +9,7 @@ import "google/protobuf/timestamp.proto";
 // The Init service definition.
 service Init {
   rpc CopyOut(CopyOutRequest) returns (stream StreamingData) {}
+  rpc LS(LSRequest) returns (stream FileInfo) {}
   rpc Reboot(google.protobuf.Empty) returns (RebootReply) {}
   rpc Shutdown(google.protobuf.Empty) returns (ShutdownReply) {}
   rpc Upgrade(UpgradeRequest) returns (UpgradeReply) {}
@@ -72,4 +73,39 @@ message StreamingData {
 message CopyOutRequest {
   // Root path to start copying data out, it might be either a file or directory
   string root_path = 1;
+}
+
+// LSRequest describes a request to list the contents of a directory
+message LSRequest {
+
+  // Root indicates the root directory for the list.  If not indicated, '/' is presumed.
+  string root = 1;
+
+  // Recurse indicates that subdirectories should be recursed.
+  bool recurse = 2;
+
+  // RecursionDepth indicates how many levels of subdirectories should be recursed.  The default (0) indicates that no limit should be enforced.
+  int32 recursion_depth = 3;
+}
+
+// FileInfo describes a file or directory's information
+message FileInfo {
+
+  // Name is the name (including prefixed path) of the file or directory
+  string name = 1;
+
+  // Size indicates the number of bytes contained within the file
+  int64 size = 2;
+
+  // Mode is the bitmap of UNIX mode/permission flags of the file
+  uint32 mode = 3;
+
+  // Modified indicates the UNIX timestamp at which the file was last modified
+  int64 modified = 4; // TODO: unix timestamp or include proto's Date type
+
+  // IsDir indicates that the file is a directory
+  bool is_dir = 5;
+
+  // Error describes any error encountered while trying to read the file information.
+  string error = 6;
 }
